Reorganized files into portfolio and windows-theme
