+++ src/components/portfolio/pages/projects/Projects.tsx
@@ -1,7 +1,7 @@
 import Fade from '@mui/material/Fade';
 import OpenInNewIcon from '@mui/icons-material/OpenInNew';
 import Technologies from '../../technologies/Technologies';
-import ProjectModel from '../../../models/projects/project';
+import ProjectModel from '../../../../models/projects/project';
 
 import styles from './Projects.module.css';
