+++ src/pages/_app.tsx
@@ -7,14 +7,14 @@ import 'bootstrap/dist/css/bootstrap.css';
 import '../styles/globals.css';
 
 import Data from '../../data.json';
-import Layout from '../components/layout/Layout';
-import TabPanel from '../components/tab-panel/TabPanel';
-import Home from '../components/pages/home/Home';
-import About from '../components/pages/about/About';
-import Projects from '../components/pages/projects/Projects';
-import Contact from '../components/pages/contact/Contact';
+import Layout from '../components/portfolio/layout/Layout';
+import TabPanel from '../components/portfolio/tab-panel/TabPanel';
+import Home from '../components/portfolio/pages/home/Home';
+import About from '../components/portfolio/pages/about/About';
+import Projects from '../components/portfolio/pages/projects/Projects';
+import Contact from '../components/portfolio/pages/contact/Contact';
 import Desktop from '../components/windows-theme/desktop/Desktop';
-import Footer from '../components/footer/Footer';
+import Footer from '../components/portfolio/footer/Footer';
 import Window from '../components/windows-theme/window/Window';
 import PageModel from '../models/component-helpers/page';
 import MenuItemModel from '../models/component-helpers/menu-item';
